requirements.txt
marshmallow>=3
flask~=1.1.2
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
flask
uwsgi
psycopg2
flask_cors
SQLAlchemy~=1.3.23
alembic~=1.5.3
passlib==1.7.4
pyjwt==2.3.0
flask_jwt_extended
